pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["KLYMENKORUS <klimenkoruslan11@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
alembic = "^1.12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
asyncpg = "^0.28.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pydantic-settings = "^2.0.3"
validators = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
